finish

+++ disney-project/src/main/java/disney/repo/ICarteRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import disney.classe.Carte;
 
-public interface ICarteRepo extends JpaRepository<Carte,Integer> {
+public interface ICarteRepo extends JpaRepository<Carte,Long> {
 
 }
